supporting old review page for v0 quest chains
