+++ package.json
@@ -18,7 +18,7 @@
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
     "@gnosis.pm/safe-apps-web3modal": "^17.0.0",
-    "@quest-chains/sdk": "^0.1.2",
+    "@quest-chains/sdk": "^0.1.4",
     "@react-three/drei": "9.26.1",
     "@react-three/fiber": "8.7.1",
     "@superfluid-finance/sdk-core": "^0.5.5",
